+++ pkg/mimir/modules.go
@@ -148,7 +148,7 @@ func newDefaultConfig() *Config {
 func (t *Mimir) initAPI() (services.Service, error) {
 	t.Cfg.API.ServerPrefix = t.Cfg.Server.PathPrefix
 
-	a, err := api.New(t.Cfg.API, t.Cfg.TenantFederation, t.Cfg.Server, t.Server, util_log.Logger)
+	a, err := api.New(t.Cfg.API, t.Cfg.TenantFederation, t.Cfg.Server, t.Server, util_log.Logger, t.ActivityTracker)
 	if err != nil {
 		return nil, err
 	}
@@ -545,7 +545,7 @@ func (t *Mimir) initQueryable() (serv services.Service, err error) {
 		t.Cfg.Querier,
 		t.Overrides,
 		t.Distributor,
-		t.AdditionalStorageQueryables,
+		t.StoreQueryable,
 		registerer,
 		util_log.Logger,
 		t.ActivityTracker,
@@ -672,10 +672,6 @@ func (t *Mimir) initQuerier() (serv services.Service, err error) {
 		&querierapi.ConsistencyExtractor{},
 	)
 
-	if t.Cfg.Querier.FilterQueryablesEnabled {
-		t.Extractors = append(t.Extractors, &querier.FilterQueryablesExtractor{})
-	}
-
 	extractor := &propagation.MultiExtractor{Extractors: t.Extractors}
 	metrics := querier.NewRequestMetrics(t.Registerer)
 	var dispatcher *querier.Dispatcher
@@ -706,7 +702,7 @@ func (t *Mimir) initQuerier() (serv services.Service, err error) {
 	// to ensure requests it processes use the default middleware instrumentation.
 	if !t.Cfg.isQuerySchedulerEnabled() && !t.Cfg.isQueryFrontendEnabled() {
 		// First, register the internal querier handler with the external HTTP server
-		t.API.RegisterQueryAPI(internalQuerierRouter, t.BuildInfoHandler)
+		t.API.RegisterQueryAPI(internalQuerierRouter, t.BuildInfoHandler, t.Cfg.Frontend.Handler.MaxBodySize)
 
 		// Second, set the http.Handler that the frontend worker will use to process requests to point to
 		// the external HTTP server. This will allow the querier to consolidate query metrics both external
@@ -745,7 +741,7 @@ func (t *Mimir) initStoreQueryable() (services.Service, error) {
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize block store queryable: %v", err)
 	}
-	t.AdditionalStorageQueryables = append(t.AdditionalStorageQueryables, querier.NewStoreGatewayTimeRangeQueryable(q, t.Cfg.Querier))
+	t.StoreQueryable = q
 	return q, nil
 }
 
@@ -951,10 +947,10 @@ func (t *Mimir) initQueryFrontend() (serv services.Service, err error) {
 	roundTripper = t.QueryFrontendTripperware(roundTripper)
 	roundTripper = querymiddleware.NewFrontendRunningRoundTripper(roundTripper, frontend, t.Cfg.Frontend.QueryMiddleware.NotRunningTimeout, util_log.Logger)
 
-	handler := transport.NewHandler(t.Cfg.Frontend.Handler, roundTripper, util_log.Logger, t.Registerer, t.ActivityTracker)
+	handler := transport.NewHandler(t.Cfg.Frontend.Handler, roundTripper, util_log.Logger, t.Registerer)
 	// Allow the Prometheus engine to be explicitly selected if MQE is in use and a fallback is configured.
 	fallbackInjector := propagation.Middleware(&streamingpromqlcompat.EngineFallbackExtractor{})
-	t.API.RegisterQueryFrontendHandler(fallbackInjector.Wrap(handler), t.BuildInfoHandler)
+	t.API.RegisterQueryFrontendHandler(fallbackInjector.Wrap(handler), t.BuildInfoHandler, t.Cfg.Frontend.Handler.MaxBodySize)
 
 	w := services.NewFailureWatcher()
 	return services.NewBasicService(func(_ context.Context) error {
@@ -1105,7 +1101,7 @@ func (t *Mimir) initRuler() (serv services.Service, err error) {
 			t.Cfg.Querier,
 			t.Overrides,
 			t.Distributor,
-			t.AdditionalStorageQueryables,
+			t.StoreQueryable,
 			rulerRegisterer,
 			util_log.Logger,
 			t.ActivityTracker,
@@ -1516,7 +1512,7 @@ func (t *Mimir) setupModuleManager() error {
 	// Add dependencies
 	deps := map[string][]string{
 		//lint:sorted
-		API:                              {Server},
+		API:                              {Server, ActivityTracker},
 		AlertManager:                     {API, MemberlistKV, Overrides, Vault},
 		BlockBuilder:                     {API, Overrides},
 		BlockBuilderScheduler:            {API},
@@ -1535,10 +1531,10 @@ func (t *Mimir) setupModuleManager() error {
 		OverridesExporter:                {Overrides, MemberlistKV, Vault},
 		Querier:                          {TenantFederation, Vault, QuerierLifecycler},
 		QuerierLifecycler:                {API, RuntimeConfig, MemberlistKV, Vault},
-		QuerierQueryPlanner:              {API, ActivityTracker, Overrides},
+		QuerierQueryPlanner:              {API, Overrides},
 		QuerierRing:                      {API, RuntimeConfig, MemberlistKV, Vault},
 		QueryFrontend:                    {QueryFrontendTripperware, MemberlistKV, Vault},
-		QueryFrontendQueryPlanner:        {API, ActivityTracker, Overrides, QuerierRing},
+		QueryFrontendQueryPlanner:        {API, Overrides, QuerierRing},
 		QueryFrontendTopicOffsetsReaders: {IngesterPartitionRing},
 		QueryFrontendTripperware:         {API, Overrides, QueryFrontendCodec, QueryFrontendTopicOffsetsReaders, QueryFrontendQueryPlanner},
 		QueryScheduler:                   {API, Overrides, MemberlistKV, Vault},