+++ operations/helm/charts/mimir-distributed/CHANGELOG.md
@@ -41,7 +41,7 @@ Entries should include a reference to the Pull Request that introduced the chang
 * [ENHANCEMENT] Default image tag to Chart.AppVersion. Use `image.tag` value to override the image tag. #13453
 * [ENHANCEMENT] Kafka: Made log retention period configurable via `kafka.logRetentionHours`. #13866
 * [ENHANCEMENT] Allow overwriting `grafana.com/min-time-between-zones-downscale` annotation value for ingester and store-gateway via `zoneAwareReplication.minTimeBetweenZonesDownscale`. #14411
-* [ENHANCEMENT] Upgrade rollout-operator chart to [0.43.0](https://github.com/grafana/helm-charts/blob/main/charts/rollout-operator). #14463
+* [ENHANCEMENT] Upgrade rollout-operator chart to [0.46.0](https://github.com/grafana/helm-charts/blob/main/charts/rollout-operator). #14463 #14854
 * [ENHANCEMENT] Add support for custom labels on PersistentVolumeClaim resources for alertmanager, compactor, ingester, and store-gateway. #14373
 * [ENHANCEMENT] Ingester: Add `ingester.zoneAwareReplication.autoIngestStorageClientRack` to pass `-ingest-storage.kafka.client-rack` when zone-aware replication is enabled. #14654
 * [ENHANCEMENT] Allow zone-aware replication for ingesters with 2 zones when ingest storage is enabled. #14449